ya crea los obetos

+++ lib/models/chole.dart
@@ -1,24 +1,24 @@
 import 'package:equatable/equatable.dart';
 
 class Chole extends Equatable {
-	final String? label;
-	final int? quantity;
-	final String? unit;
+  final String? label;
+  final double? quantity;
+  final String? unit;
 
-	const Chole({this.label, this.quantity, this.unit});
+  const Chole({this.label, this.quantity, this.unit});
 
-	factory Chole.fromJson(Map<String, dynamic> json) => Chole(
-				label: json['label'] as String?,
-				quantity: json['quantity'] as int?,
-				unit: json['unit'] as String?,
-			);
+  factory Chole.fromJson(Map<String, dynamic> json) => Chole(
+        label: json['label'] as String?,
+        quantity: json['quantity'] as double?,
+        unit: json['unit'] as String?,
+      );
 
-	Map<String, dynamic> toJson() => {
-				'label': label,
-				'quantity': quantity,
-				'unit': unit,
-			};
+  Map<String, dynamic> toJson() => {
+        'label': label,
+        'quantity': quantity,
+        'unit': unit,
+      };
 
-	@override
-	List<Object?> get props => [label, quantity, unit];
+  @override
+  List<Object?> get props => [label, quantity, unit];
 }
